refactor project structure to consolidate service and controller paths
