+++ src/logger/logger.js
@@ -14,11 +14,11 @@ const logger = winston.createLogger({
   defaultMeta: { service: "user-auth-service" },
   transports: [
     new winston.transports.File({
-      filename: path.join(__dirname, "../../../logs/error.log"),
+      filename: path.join(__dirname, "../../logs/error.log"),
       level: "error",
     }),
     new winston.transports.File({
-      filename: path.join(__dirname, "../../../logs/combined.log"),
+      filename: path.join(__dirname, "../../logs/combined.log"),
     }),
   ],
 });
