+++ src/services/jwt.service.js
@@ -1,5 +1,5 @@
 const jwt = require("jsonwebtoken");
-const config = require("../../config/config");
+const config = require("../config/config");
 
 class JWTService {
   constructor() {
