+++ src/use-cases/auth/signup.use-case.js
@@ -1,12 +1,12 @@
 class SignupUseCase {
-  constructor(userRepository, jwtService) {
-    this.userRepository = userRepository;
+  constructor(authRepository, jwtService) {
+    this.authRepository = authRepository;
     this.jwtService = jwtService;
   }
 
   async execute(userData) {
     try {
-      const existingUser = await this.userRepository.findByEmail(
+      const existingUser = await this.authRepository.findByEmail(
         userData.email
       );
 
@@ -14,7 +14,7 @@ class SignupUseCase {
         throw new Error("Email already registered");
       }
 
-      const user = await this.userRepository.create(userData);
+      const user = await this.authRepository.create(userData);
       const token = this.jwtService.generateToken(user);
 
       return { user, token };
