+++ src/webserver/express-app.js
@@ -2,7 +2,7 @@ const express = require("express");
 const cors = require("cors");
 const authRoutes = require("./routes/auth.routes");
 const userRoutes = require("./routes/user.routes");
-const globalErrorMiddleware = require("../../interfaces/middlewares/globalError.middleware");
+const globalErrorMiddleware = require("../middlewares/globalError.middleware");
 
 const app = express();
